Changed to OOP. Added multiple slider capability. Added randomize functionality
Instead of storing slides in an array, slides are stored in an object.
M
ore than one slider may be added to a page. The plugin now looks for the
 class "ac-fader" instead of the id attribute.
Slides in an individual f
ader may be randomized by adding the class "ac-randomize" to the div. Al
ternately, all sliders on a page may be randomized by changing the varia
ble ac-randomize to true in ac-fader.js.